pyproject.toml
[tool.poetry]
name = "knowledge-base-guardian"
version = "0.0.11"
description = "An LLM application to safeguard the consistency of documents in a knowledge base"
authors = ["Senne Batsleer <senne@dataroots.io>"]
license = "MIT"
readme = "README.md"
repository = "https://github.com/datarootsio/knowledgebase_guardian/"
packages = [{include = "kb_guardian"}]
include = ["default_config.yml"]

[tool.poetry.dependencies]
python = "^3.9"
langchain = "0.0.213"
faiss-cpu = "1.7.3"
python-dotenv = "1.0.0"
transformers = "4.30.2"
bs4 = "0.0.1"
openai = "^0.27.0"
tiktoken = "^0.4.0"

[tool.poetry.group.dev.dependencies]
black = "^23.3.0"
flake8 = "^6.0.0"
isort = "^5.10.1"
pytest = "^7.3.1"
pytest-mock = "^3.11.1"
pytest-cov = "^4.1.0"
pre-commit = "^3.3.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
profile = "black"
